+++ src/core/config/config.rs
@@ -15,7 +15,7 @@ use tailcall_typedefs_common::ServiceDocumentBuilder;
 use super::telemetry::Telemetry;
 use super::{KeyValue, Link, Server, Upstream};
 use crate::core::config::from_document::from_document;
-use crate::core::config::npo::Queries;
+use crate::core::config::npo::QueryPath;
 use crate::core::config::source::Source;
 use crate::core::directive::DirectiveCodec;
 use crate::core::http::Method;
@@ -873,8 +873,8 @@ impl Config {
         }
     }
 
-    pub fn n_plus_one(&self) -> Queries {
-        super::npo::Identifier::new(self).identify()
+    pub fn n_plus_one(&self) -> QueryPath {
+        super::npo::PathTracker::new(self).find()
     }
 
     ///
