+++ src/core/config/mod.rs
@@ -3,7 +3,7 @@ pub use config::*;
 pub use config_module::*;
 pub use key_values::*;
 pub use link::*;
-pub use npo::Queries;
+pub use npo::QueryPath;
 pub use reader_context::*;
 pub use server::*;
 pub use source::*;
